requirements.txt
streamlit
opendataloader-pdf
pdfplumber
pymupdf
anthropic
markitdown[all]
markitdown-ocr
python-pptx
pillow
python-dotenv
pyyaml
reportlab
